fix: tekton ci cel condition and remove script

+++ scripts/remove-dependencies-pre.sh
@@ -1,5 +1,7 @@
 # Remove keda CR
-oc delete --ignore-not-found kedacontroller keda -n openshift-keda
+if oc get crd kedacontrollers.keda.sh &>/dev/null; then
+  oc delete --ignore-not-found kedacontroller keda -n openshift-keda
+fi
 
 # Remove NFD CR
 if oc get crd nodefeaturediscoveries.nfd.openshift.io &>/dev/null; then
